use context path of resource handler from a request parameter
Removed the student configuration property and replaced it with a parameter on the rest endpoint.
This allows the service to be called from both proctor and student web applications.
The context path parameter is used to construct a relative url to retrieve additional resources to display from an item on an exam.